find apps for all other platforms also
